Custom-Search-Engine/requirements.txt
# requirements.txt
Flask
mysql-connector-python
requests
